#11 Refactor Home Out of Nero
